requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.4
    # via web3
aiosignal==1.3.1
    # via aiohttp
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via django
asttokens==2.2.1
    # via stack-data
async-timeout==4.0.2
    # via
    #   aiohttp
    #   redis
attrs==23.1.0
    # via
    #   aiohttp
    #   jsonschema
backcall==0.2.0
    # via ipython
base58==2.1.1
    # via multiaddr
billiard==4.1.0
    # via celery
bitarray==2.7.5
    # via eth-account
boto3==1.26.133
    # via -r requirements.in
botocore==1.29.133
    # via
    #   boto3
    #   s3transfer
build==0.10.0
    # via pip-tools
celery==5.3.1
    # via
    #   -r requirements.in
    #   django-celery-beat
certifi==2022.12.7
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   pip-tools
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cron-descriptor==1.2.35
    # via django-celery-beat
cryptography==41.0.0
    # via pyopenssl
cytoolz==0.12.1
    # via
    #   eth-keyfile
    #   eth-utils
decorator==5.1.1
    # via ipython
django==4.2.3
    # via
    #   -r requirements.in
    #   django-bulk-load
    #   django-celery-beat
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-extensions
    #   django-model-utils
    #   django-redis
    #   djangorestframework
django-autoslug==1.9.9
    # via -r requirements.in
django-bulk-load==1.2.2
    # via -r requirements.in
django-celery-beat==2.5.0
    # via -r requirements.in
django-cors-headers==4.0.0
    # via -r requirements.in
django-debug-toolbar==4.1.0
    # via -r requirements.in
django-environ==0.10.0
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-model-utils==4.3.1
    # via -r requirements.in
django-redis==5.3.0
    # via -r requirements.in
django-timezone-field==5.0
    # via django-celery-beat
djangorestframework==3.14.0
    # via -r requirements.in
eth-abi==2.2.0
    # via
    #   eth-account
    #   web3
eth-account==0.5.9
    # via web3
eth-hash[pycryptodome]==0.5.2
    # via
    #   eth-utils
    #   web3
eth-keyfile==0.5.1
    # via eth-account
eth-keys==0.3.4
    # via
    #   eth-account
    #   eth-keyfile
eth-retry==0.1.18
    # via multicall
eth-rlp==0.2.1
    # via eth-account
eth-typing==2.3.0
    # via
    #   eth-abi
    #   eth-keys
    #   eth-utils
    #   web3
eth-utils==1.9.5
    # via
    #   eth-abi
    #   eth-account
    #   eth-keyfile
    #   eth-keys
    #   eth-rlp
    #   rlp
    #   web3
executing==1.2.0
    # via stack-data
factory-boy==3.2.1
    # via -r requirements.in
faker==18.11.1
    # via factory-boy
freezegun==1.2.2
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
gunicorn==20.1.0
    # via -r requirements.in
hexbytes==0.3.1
    # via
    #   eth-account
    #   eth-rlp
    #   web3
hiredis==2.2.3
    # via -r requirements.in
idna==3.4
    # via
    #   requests
    #   yarl
ijson==3.2.0.post0
    # via -r requirements.in
iniconfig==2.0.0
    # via pytest
ipfshttpclient==0.8.0a2
    # via web3
ipython==8.14.0
    # via -r requirements.in
jedi==0.18.2
    # via ipython
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==3.2.0
    # via web3
kombu==5.3.1
    # via celery
lru-dict==1.2.0
    # via web3
markupsafe==2.1.3
    # via werkzeug
matplotlib-inline==0.1.6
    # via ipython
multiaddr==0.0.9
    # via ipfshttpclient
multicall==0.6.2
    # via -r requirements.in
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
netaddr==0.8.0
    # via multiaddr
packaging==23.1
    # via
    #   -r requirements.in
    #   build
    #   pytest
    #   pytest-sugar
parsimonious==0.8.1
    # via eth-abi
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.13.0
    # via -r requirements.in
pluggy==1.2.0
    # via pytest
prompt-toolkit==3.0.38
    # via
    #   click-repl
    #   ipython
protobuf==3.19.5
    # via web3
psycopg2==2.9.6
    # via
    #   -r requirements.in
    #   django-bulk-load
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pycryptodome==3.18.0
    # via
    #   eth-hash
    #   eth-keyfile
pygments==2.14.0
    # via ipython
pyopenssl==23.2.0
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
pyrsistent==0.19.3
    # via jsonschema
pytest==7.3.1
    # via
    #   -r requirements.in
    #   pytest-django
    #   pytest-mock
    #   pytest-sugar
pytest-django==4.5.2
    # via -r requirements.in
pytest-mock==3.10.0
    # via -r requirements.in
pytest-sugar==0.9.6
    # via -r requirements.in
python-crontab==2.7.1
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   botocore
    #   celery
    #   faker
    #   freezegun
    #   python-crontab
pytz==2022.7.1
    # via
    #   django-timezone-field
    #   djangorestframework
pyyaml==6.0
    # via responses
redis==4.5.4
    # via
    #   -r requirements.in
    #   django-redis
requests==2.31.0
    # via
    #   -r requirements.in
    #   ipfshttpclient
    #   responses
    #   web3
responses==0.23.1
    # via -r requirements.in
rlp==2.0.1
    # via
    #   eth-account
    #   eth-rlp
s3transfer==0.6.0
    # via boto3
sentry-sdk==1.22.1
    # via -r requirements.in
six==1.16.0
    # via
    #   asttokens
    #   click-repl
    #   jsonschema
    #   multiaddr
    #   parsimonious
    #   python-dateutil
sqlparse==0.4.4
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.2
    # via ipython
statsd==4.0.1
    # via -r requirements.in
termcolor==2.3.0
    # via pytest-sugar
toolz==0.12.0
    # via cytoolz
traitlets==5.9.0
    # via
    #   ipython
    #   matplotlib-inline
types-pyyaml==6.0.12.10
    # via responses
typing-extensions==4.6.3
    # via eth-retry
tzdata==2022.7
    # via
    #   celery
    #   django-celery-beat
urllib3==1.26.15
    # via
    #   botocore
    #   requests
    #   responses
    #   sentry-sdk
varint==1.0.2
    # via multiaddr
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
web3==5.28.0
    # via
    #   -r requirements.in
    #   multicall
websockets==9.1
    # via web3
werkzeug==2.3.6
    # via -r requirements.in
wheel==0.40.0
    # via pip-tools
whitenoise==6.5.0
    # via -r requirements.in
yarl==1.9.2
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
